# chat_server/api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "chat-server",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/chats": {
      "get": {
        "tags": [
          "chat"
        ],
        "operationId": "list_chat_handler",
        "responses": {
          "200": {
            "description": "List of chats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Chat"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "token": []
          }
        ]
      },
      "post": {
        "tags": [
          "chat"
        ],
        "operationId": "create_chat_handler",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChat"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Chat created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chat"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": []
          }
        ]
      }
    },
    "/api/chats/{id}": {
      "get": {
        "tags": [
          "chat"
        ],
        "operationId": "get_chat_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Chat id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chat found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chat"
                }
              }
            }
          },
          "404": {
            "description": "Chat not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOutput"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": []
          }
        ]
      }
    },
    "/api/signin": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "signin_handler",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SigninUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthOutput"
                }
              }
            }
          }
        }
      }
    },
    "/api/signup": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "signup_handler",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthOutput"
                }
              }
            }
          }
        }
      }
    },
    "/api/upload": {
      "post": {
        "tags": [
          "message"
        ],
        "operationId": "upload_handler",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/FileForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Chat file error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOutput"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AuthOutput": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "Chat": {
        "type": "object",
        "required": [
          "id",
          "ws_id",
          "type",
          "members",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "members": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/ChatType"
          },
          "ws_id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ChatType": {
        "type": "string",
        "enum": [
          "single",
          "group",
          "private_channel",
          "public_channel"
        ]
      },
      "CreateUser": {
        "type": "object",
        "required": [
          "fullname",
          "email",
          "password",
          "workspace"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "fullname": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "workspace": {
            "type": "string"
          }
        }
      },
      "ErrorOutput": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "FileForm": {
        "type": "object",
        "required": [
          "file"
        ],
        "properties": {
          "file": {
            "type": "string",
            "format": "binary"
          }
        }
      },
      "SigninUser": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "token": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "tags": [
    {
      "name": "chat",
      "description": "Chat related operations"
    }
  ]
}